+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Changed
 - Add more protection against bad stats options
 - Upgrade docker image to python 3.11
 - Update docker-compose file to use postgres 16
+- Use SQLAlchemy-Utils to make new db
 
 Fixed
 -----